Update extension skeleton with solidus_dev_support
