+++ .github_changelog_generator
@@ -1,0 +1,2 @@
+issues=false
+exclude-labels=infrastructure
