+++ .rubocop.yml
@@ -4,7 +4,7 @@ require:
   - solidus_dev_support/rubocop
 
 AllCops:
-  NewCops: enable
+  NewCops: disable
 
 Layout/LineLength:
   Enabled: false
