+++ app/assets/javascripts/spree/backend/solidus_shipstation.js
@@ -1,0 +1,2 @@
+// Placeholder manifest file.
+// the installer will append this file to the app vendored assets here: vendor/assets/javascripts/spree/backend/all.js'
