+++ app/assets/stylesheets/spree/backend/solidus_shipstation.css
@@ -1,0 +1,4 @@
+/*
+Placeholder manifest file.
+the installer will append this file to the app vendored assets here: 'vendor/assets/stylesheets/spree/backend/all.css'
+*/
