+++ app/assets/stylesheets/spree/frontend/solidus_shipstation.css
@@ -1,0 +1,4 @@
+/*
+Placeholder manifest file.
+the installer will append this file to the app vendored assets here: 'vendor/assets/stylesheets/spree/frontend/all.css'
+*/
