+++ bin/sandbox
@@ -67,12 +67,12 @@ unbundled bundle install --gemfile Gemfile
 
 unbundled bundle exec rake db:drop db:create
 
-unbundled bundle exec rails generate spree:install \
+unbundled bundle exec rails generate solidus:install \
   --auto-accept \
   --user_class=Spree::User \
   --enforce_available_locales=true \
   --with-authentication=false \
-  --payment-method=none
+  --payment-method=none \
   $@
 
 unbundled bundle exec rails generate solidus:auth:install
