+++ lib/generators/solidus_shipstation/install/templates/initializer.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+SolidusShipstation.configure do |config|
+  # TODO: Remember to change this with the actual preferences you have implemented!
+  # config.sample_preference = 'sample_value'
+end
