+++ lib/solidus_shipstation/configuration.rb
@@ -5,4 +5,16 @@ module SolidusShipstation
     attr_accessor :username, :password, :weight_units, :ssl_encrypted, :capture_at_notification,
       :export_canceled_shipments
   end
+
+  class << self
+    def configuration
+      @configuration ||= Configuration.new
+    end
+
+    alias config configuration
+
+    def configure
+      yield configuration
+    end
+  end
 end
