+++ lib/solidus_shipstation/testing_support/factories.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+FactoryBot.define do
+end
