+++ solidus_shipstation.gemspec
@@ -15,6 +15,7 @@ Gem::Specification.new do |spec|
 
   spec.metadata['homepage_uri'] = spec.homepage
   spec.metadata['source_code_uri'] = 'https://github.com/solidusio-contrib/solidus_shipstation'
+  spec.metadata['changelog_uri'] = 'https://github.com/solidusio-contrib/solidus_shipstation/blob/master/CHANGELOG.md'
 
   spec.required_ruby_version = Gem::Requirement.new('~> 2.5')
 
@@ -33,5 +34,5 @@ Gem::Specification.new do |spec|
 
   spec.add_development_dependency 'rails-controller-testing'
   spec.add_development_dependency 'rspec-xsd'
-  spec.add_development_dependency 'solidus_dev_support'
+  spec.add_development_dependency 'solidus_dev_support', '~> 2.4'
 end
