+++ spec/spec_helper.rb
@@ -18,8 +18,8 @@ require 'solidus_dev_support/rspec/feature_helper'
 # in spec/support/ and its subdirectories.
 Dir["#{__dir__}/support/**/*.rb"].sort.each { |f| require f }
 
-# Requires factories defined in lib/solidus_shipstation/factories.rb
-require 'solidus_shipstation/factories'
+# Requires factories defined in lib/solidus_shipstation/testing_support/factories.rb
+SolidusDevSupport::TestingSupport::Factories.load_for(SolidusShipstation::Engine)
 
 RSpec.configure do |config|
   config.infer_spec_type_from_file_location!
